* Fixed network driver: Added the AUTOINC flag and did minor rewriting  * Changed clock a bit to be closer to reality

+++ trunk/project/src/trunk/arch/lpc21xx/clock/clock-arch.c
@@ -41,7 +41,7 @@
 #include "clock-arch.h"
 #include "type.h"
 
-uint32_t tick;
+uint64_t tick;
 
 
 void clock_init(void);
